insights_mcp: move MCPs into a separate file to avoid circular import
