+++ src/insights_mcp/server.py
@@ -15,31 +15,11 @@ import requests
 from fastmcp import FastMCP
 from mcp.types import Icon, ToolAnnotations
 
-from advisor_mcp.server import mcp_server as AdvisorMCP
-from content_sources_mcp.server import mcp as ContentSourcesMCP
-from image_builder_mcp.server import mcp_server as ImageBuilderMCP
 from insights_mcp import __version__, config
 from insights_mcp.catalog_tools import catalog_tool_description
 from insights_mcp.mcp import InsightsMCP
 from insights_mcp.oauth import create_oauth_provider
-from inventory_mcp.server import mcp as InventoryMCP
-from planning_mcp.server import mcp as PlanningMCP
-from rbac_mcp.server import mcp as RbacMCP
-from remediations_mcp.server import mcp as RemediationsMCP
-from rhsm_mcp.server import mcp as RhsmMCP
-from vulnerability_mcp.server import mcp as VulnerabilityMCP
-
-MCPS: list[InsightsMCP] = [
-    ImageBuilderMCP,
-    RhsmMCP,
-    VulnerabilityMCP,
-    RemediationsMCP,
-    AdvisorMCP,
-    InventoryMCP,
-    ContentSourcesMCP,
-    RbacMCP,
-    PlanningMCP,
-]
+from insights_mcp.toolsets import MCPS
 
 
 def _format_server_tools(
